Module strcursor::grapheme

Defines types for representing single grapheme clusters.

Structs

Gc
A slice of a single Unicode grapheme cluster (GC) (akin to str).

GcBuf
An owned, single Unicode grapheme cluster (akin to String).

Type Definitions

ToLowercase
An iterator over the lower case mapping of a given grapheme cluster, returned from Gc::to_lowercase.

ToUppercase
An iterator over the lower case mapping of a given grapheme cluster, returned from Gc::to_uppercase.